fix(practice): fix layout issue; remove /wajo directory; render latex
